+++ primefaces/src/main/java/org/primefaces/component/staticmessage/StaticMessageBase.java
@@ -23,22 +23,31 @@
  */
 package org.primefaces.component.staticmessage;
 
+import java.util.Collection;
 import javax.faces.component.UIComponentBase;
+import javax.faces.component.behavior.ClientBehaviorHolder;
+import org.primefaces.component.api.MixedClientBehaviorHolder;
+import org.primefaces.component.api.Widget;
+import org.primefaces.util.LangUtils;
 
-public abstract class StaticMessageBase extends UIComponentBase {
+public abstract class StaticMessageBase extends UIComponentBase implements Widget, ClientBehaviorHolder, MixedClientBehaviorHolder {
 
     public static final String COMPONENT_FAMILY = "org.primefaces.component";
 
     public static final String DEFAULT_RENDERER = "org.primefaces.component.StaticMessageRenderer";
 
+    protected static final Collection<String> EVENT_NAMES = LangUtils.unmodifiableList("close");
+
     public enum PropertyKeys {
+        widgetVar,
         summary,
         detail,
         display,
         escape,
         style,
         styleClass,
-        severity;
+        severity,
+        closable
     }
 
     public StaticMessageBase() {
@@ -48,6 +57,14 @@ public abstract class StaticMessageBase extends UIComponentBase {
     @Override
     public String getFamily() {
         return COMPONENT_FAMILY;
+    }
+
+    public String getWidgetVar() {
+        return (String) getStateHelper().eval(PropertyKeys.widgetVar, null);
+    }
+
+    public void setWidgetVar(final String widgetVar) {
+        getStateHelper().put(PropertyKeys.widgetVar, widgetVar);
     }
 
     public boolean isEscape() {
@@ -104,5 +121,13 @@ public abstract class StaticMessageBase extends UIComponentBase {
 
     public void setStyleClass(String styleClass) {
         getStateHelper().put(PropertyKeys.styleClass, styleClass);
+    }
+
+    public boolean isClosable() {
+        return (Boolean) getStateHelper().eval(PropertyKeys.closable, false);
+    }
+
+    public void setClosable(boolean closable) {
+        getStateHelper().put(PropertyKeys.closable, closable);
     }
 }