+++ docs/14_0_0/gettingstarted/whatsnew.md
@@ -38,6 +38,9 @@ Look into [migration guide](https://primefaces.github.io/primefaces/14_0_0/#/../
 * SelectCheckBoxMenu
     * Added `selectedLabel` to display a label when items selected similar to `emptyMessage`.
     
+* StaticMessage
+    * Added `closable` to add a close icon to hide the message similar to `Messages`.
+
 * TieredMenu
     * Added `showDelay` delay in milliseconds before displaying the submenu. Default is 0 meaning immediate.
     * Added `hideDelay` delay in milliseconds before hiding the submenu., if 0 not hidden until document.click(). Default is 0.